hard-float support, fixed kernel scheduling to support multi-core

+++ dogfood/vecadd2/kernel.cl
@@ -1,7 +1,7 @@
 kernel void
-vecadd (__global const int *a,
-	__global const int *b,
-	__global int *c)
+vecadd (__global const float *a,
+	__global const float *b,
+	__global float *c)
 {
   int gid = get_global_id(0);
   c[gid] = a[gid] + b[gid];
